requirements.txt
chainlit>=1.0.0
openai-agents>=0.0.6
openai>=1.68.0
pydantic>=2.0.0
python-dotenv>=1.0.0
